+++ backend/internal/db/mongo/transaction_crud.go
@@ -184,7 +184,7 @@ func (b *Backend) GetDeletedTransactions(page uint64, size uint64) ([]*models.Tr
 	return accs, nil
 }
 
-func (b *Backend) CountDeletedTransactions() (int64, error) {
+func (b *Backend) CountDeletedTransactions() (uint64, error) {
 	ctx, cancel := b.GetContext()
 	defer cancel()
 
@@ -197,5 +197,5 @@ func (b *Backend) CountDeletedTransactions() (int64, error) {
 		return 0, err
 	}
 
-	return count, nil
+	return uint64(count), nil
 }
